requirements.txt
absl-py==0.11.0
asgiref==3.3.4
astor==0.8.1
astunparse==1.6.3
bert-serving==0.0.1
bert-serving-client==1.10.0
bert-serving-server==1.10.0
cached-property==1.5.2
cachetools==4.2.0
category-encoders==2.2.2
certifi==2020.12.5
cffi==1.14.2
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.8
cycler==0.10.0
decorator==4.4.2
Django==3.2.4
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.2.2
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
GPUtil==1.4.0
grpcio==1.32.0
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.3.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.4.2
mkl==2021.2.0
mkl-include==2021.2.0
mock==4.0.3
networkx==2.5.1
numpy==1.19.4
oauthlib==3.1.0
opt-einsum==3.3.0
pandas==1.2.2
patsy==0.5.1
Pillow==8.2.0
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyOpenSSL==20.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==20.0.0
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.6
scikit-image==0.18.1
scikit-learn==0.24.1
scipy==1.6.0
seaborn==0.11.1
six==1.15.0
sklearn==0.0
Spidy==0.2
sqlparse==0.4.1
statsmodels==0.12.2
tbb==2021.2.0
tensorboard==1.15.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torchvision==0.9.1
tqdm==4.60.0
typing==3.7.4.3
typing-extensions==3.7.4.3
urllib3==1.26.2
Werkzeug==2.0.1
wrapt==1.12.1
zipp==3.4.0
